pyproject.toml
[tool.poetry]
name = "pythonmath"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sympy = "^1.11.1"
numpy = "^1.24.0"
matplotlib = "^3.6.2"
scipy = "^1.9.3"
jupyter = "^1.0.0"
plotly = "^5.12.0"
networkx = "^3.0"
pandas = "^1.5.3"
neo4j-driver = "^5.6.0"
seaborn = "^0.12.2"
dask = "^2023.3.1"
pulp = "^2.7.0"
mypulp = "^0.0.11"
bokeh = "^3.1.0"
blaze = "^0.10.1"
statsmodels = "^0.13.5"
scop = "^0.0.1"
ipython = "^8.11.0"
pyglet = "^2.0.5"
coq-jupyter = "^1.6.0"
python-barcode = {extras = ["images"], version = "^0.15.1"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.0.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.1.0"
flake8-isort = "^6.0.0"
pyproject-flake8 = "^6.0.0.post1"
flake8-bugbear = "^23.2.13"
flake8-builtins = "^2.1.0"
flake8-unused-arguments = "^0.0.13"
pep8-naming = "^0.13.3"
flake8-pytest-style = "^1.7.2"
bandit = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.flake8]
max-line-length = 79
# E203: ":"の前の空白を入れないルール
# W503: 演算子の前に改行しないようにするルール
extend-ignore = ["E203", "W503"]
exclude = [".venv", ".git", "__pycache__",]
max-complexity = 10

[tool.isort]
profile = "black"
line_length = 79
# 各ライブラリ群の説明を追記する
import_heading_stdlib      = "Standard Library"
import_heading_thirdparty  = "Third Party Library"
import_heading_firstparty  = "First Party Library"
import_heading_localfolder = "Local Library"
# from third_party import lib1, lib2...のような記述時の改行方法の設定(https://pycqa.github.io/isort/docs/configuration/multi_line_output_modes.html)
multi_line_output = 3
# 最後の要素の末尾に","を付けるようにする設定
include_trailing_comma = true

[tool.mypy]
# エラー時のメッセージを詳細表示
show_error_context = true
# エラー発生箇所の行数/列数を表示
show_column_numbers = true
# import 先のチェックを行わない (デフォルトだとサードパーティーライブラリまでチェックする)
ignore_missing_imports = true
# 関数定義の引数/戻り値に型アノテーション必須
disallow_untyped_defs = true
# デフォルト引数に None を取る場合型アノテーションに Optional 必須
no_implicit_optional = true
# 戻り値が Any 型ではない関数の戻り値の型アノテーションが Any のとき警告
warn_return_any = true
# mypy エラーに該当しない箇所に `# type: ignore` コメントが付与されていたら警告
# ※ `# type: ignore` が付与されている箇所は mypy のエラーを無視出来る
warn_unused_ignores = true
# 冗長なキャストに警告
warn_redundant_casts = true